requirements.txt
plotly>=6.0.0
networkx>=3.0
folium>=0.14.0
requests>=2.28.0
pillow>=9.0.0
